Move vectorized sampler from tf to garage.sampler
- refactor tests accordingly
- fix tf_vars initialize bug in examples/np
