+++ src/garage/np/algos/off_policy_rl_algorithm.py
@@ -1,5 +1,5 @@
 from garage.np.algos import RLAlgorithm
-from garage.tf.samplers import OffPolicyVectorizedSampler
+from garage.sampler import OffPolicyVectorizedSampler
 
 
 class OffPolicyRLAlgorithm(RLAlgorithm):
