+++ src/garage/sampler/__init__.py
@@ -2,6 +2,10 @@ from garage.sampler.base import BaseSampler
 from garage.sampler.base import Sampler
 from garage.sampler.batch_sampler import BatchSampler
 from garage.sampler.is_sampler import ISSampler
+from garage.sampler.off_policy_vectorized_sampler import (
+    OffPolicyVectorizedSampler)
+from garage.sampler.on_policy_vectorized_sampler import (
+    OnPolicyVectorizedSampler)
 from garage.sampler.parallel_vec_env_executor import ParallelVecEnvExecutor
 from garage.sampler.ray_sampler import RaySampler, SamplerWorker
 from garage.sampler.stateful_pool import singleton_pool
@@ -9,5 +13,6 @@ from garage.sampler.vec_env_executor import VecEnvExecutor
 
 __all__ = [
     'BaseSampler', 'BatchSampler', 'Sampler', 'ISSampler', 'singleton_pool',
-    'RaySampler', 'SamplerWorker', 'ParallelVecEnvExecutor', 'VecEnvExecutor'
+    'RaySampler', 'SamplerWorker', 'ParallelVecEnvExecutor', 'VecEnvExecutor',
+    'OffPolicyVectorizedSampler', 'OnPolicyVectorizedSampler'
 ]
