+++ src/garage/tf/samplers/__init__.py
@@ -1,11 +1,4 @@
 from garage.tf.samplers.batch_sampler import BatchSampler
-from garage.tf.samplers.off_policy_vectorized_sampler import (
-    OffPolicyVectorizedSampler)
-from garage.tf.samplers.on_policy_vectorized_sampler import (
-    OnPolicyVectorizedSampler)
 from garage.tf.samplers.ray_sampler import (RaySamplerTF, SamplerWorkerTF)
 
-__all__ = [
-    'BatchSampler', 'OffPolicyVectorizedSampler', 'OnPolicyVectorizedSampler',
-    'RaySamplerTF', 'SamplerWorkerTF'
-]
+__all__ = ['BatchSampler', 'RaySamplerTF', 'SamplerWorkerTF']
