+++ tests/garage/np/algos/test_cem.py
@@ -2,10 +2,10 @@ import pytest
 
 from garage.np.algos import CEM
 from garage.np.baselines import LinearFeatureBaseline
+from garage.sampler import OnPolicyVectorizedSampler
 from garage.tf.envs import TfEnv
 from garage.tf.experiment import LocalTFRunner
 from garage.tf.policies import CategoricalMLPPolicy
-from garage.tf.samplers import OnPolicyVectorizedSampler
 from tests.fixtures import TfGraphTestCase
