+++ tests/garage/sampler/test_ray_batched_sampler.py
@@ -5,10 +5,8 @@ import ray
 
 from garage.envs.grid_world_env import GridWorldEnv
 from garage.np.policies import ScriptedPolicy
-from garage.sampler import RaySampler, SamplerWorker
+from garage.sampler import OnPolicyVectorizedSampler, RaySampler, SamplerWorker
 from garage.tf.envs import TfEnv
-from garage.tf.samplers.on_policy_vectorized_sampler \
-    import OnPolicyVectorizedSampler
 
 
 class TestSampler:
